Update web assets build system
