+++ web/src/slide/projector_current_list_of_speakers.css
