+++ web/src/slide/projector_current_speaker_chyron.css
